# back-furniture/src/api/swagger.yaml
openapi: 3.0.0
info:
  title: Furniture Store
  description: This is Furniture Shop Backend 
  version: 0.1.9
servers:
  - url: /api/v1
security:
  - BearerAuth: []
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: Enter your Bearer token here
consumes:
  - application/json 
paths:
...
  /productHome:
    get:
      tags:
        - "ProductHome"
      x-swagger-router-controller: productHome
      operationId: getProductHome
      summary: Returns a list of products.
      parameters:
        - name: currentPage
          in: query
          description: currentPage
          required: false
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          description: pageSize
          required: false
          schema:
            type: integer
            minimum: 1
        - name: sort
          in: query
          description: "sort column"
          required: false
          schema:
            type: integer
        - name: sortDirection
          in: query
          description: "sort i.e asc desc"
          required: false
          schema:
            type: integer
            enum: 
              - asc
              - desc                                   
      responses:
        '200':    # status code
          description: Default Response
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/ProductHomes"
    post: 
      x-swagger-router-controller: productHome
      operationId: addProductHome
      summary: Register a new productHome.
      tags:
        - "ProductHome"      
      description: Register a new productHome
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                productName:
                  type: string
                  example: Door
                imageUrl:
                  type: string
                  example: sumit
                description:
                  type: string
                  example: good quality
                productCategory:
                  type: string
                  example: Doors
      responses: 
        '200':
          description: Created 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductHome"      
    delete:
      tags:
        - "ProductHome"
      x-swagger-router-controller: producthome
      operationId: deleteAllProductHome
      summary: Delete All Available producthome.
      responses:
        '204':    # status code
          description: Default Response
  /productHome/{productHomeId}:
    get:
      tags:
        - "ProductHome"
      x-swagger-router-controller: productHome
      operationId: getProductHomeById
      summary: Returns  productHome.
      parameters:
        - name: productHomeId
          in: path
          description: productHomeId
          required: false
          schema:
            type: string
      responses:
        '200':    # status code
          description: Default Response
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/ProductHome"
    patch: 
      x-swagger-router-controller: productHome
      operationId: updateProductHomeById
      summary: update in a productHome.
      tags:
        - "ProductHome"      
      description: Register a new productHome
      parameters:
        - name: productHomeId
          in: path
          description: productHomeId
          required: false
          schema:
            type: string      
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                productName:
                  type: string
                  example: Wooden door
                imageUrl:
                  type: string
                  example: door.jpg
                description:
                  type: string
                  example: Good quality
                productCategory:
                  type: string
                  example: doors
      responses: 
        '200':
          description: Created 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductHome"      
    delete:
      tags:
        - "ProductHome"
      x-swagger-router-controller: productHome
      operationId: deleteProductHomeById
      summary: Delete All Available productHome.
      parameters:
        - name: productHomeId
          in: path
          description: productHomeId
          required: false
          schema:
            type: string        
      responses:
        '204':    # status code
          description: Default Response           
...
  /productDetail:
    get:
      tags:
        - "ProductDetail"
      x-swagger-router-controller: productDetail
      operationId: getProductDetail
      summary: Returns a list of productDetail.
      parameters:
        - name: currentPage
          in: query
          description: currentPage
          required: false
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          description: pageSize
          required: false
          schema:
            type: integer
            minimum: 1
        - name: sort
          in: query
          description: "sort column"
          required: false
          schema:
            type: integer
        - name: sortDirection
          in: query
          description: "sort i.e asc desc"
          required: false
          schema:
            type: integer
            enum: 
              - asc
              - desc                                   
      responses:
        '200':    # status code
          description: Default Response
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/ProductDetails"
    post: 
      x-swagger-router-controller: productDetail
      operationId: addProductDetail
      summary: Register a new productDetail.
      tags:
        - "ProductDetail"      
      description: Register a new productDetail
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                productName:
                  type: string
                  example: Door
                productId:
                  type: string
                  example: sumit
                description:
                  type: string
                  example: goodquality
                category:
                  type: string
                  example: Doors
                price:
                  type: number
                  example: 233
                homeCategory:
                  type: string
                  example: bed
                subcategory:
                  type: string
                  example: goodquality
                availability:
                  type: boolean
                  example: true
                quantity:
                  type: number
                  example: 2
                rating:
                  type: number
                  example: 4
                images:
                  type: array
                  items:
                    type: string
                  example:
                    - bed
                    - sofa 
                assembly:
                  type: string
                  example: anything
                brand:
                  type: string
                  example: SARA
                material:
                  type: string
                  example: wooden
                color:
                  type: string
                  example: brown
                type:
                  type: string
                  example: bed
                totalOrder:
                  type: number
                  example: 5
                dimensionInInch:
                  type: string
                  example: 12in   
                dimensionInCem:
                  type: string
                  example: 2cm
                storage:
                  type: string
                  example: xyz
                weight:
                  type: string
                  example: 40KG.
                warranty:
                  type: number
                  example: 2                                     
      responses: 
        '200':
          description: Created 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"      
    delete:
      tags:
        - "ProductDetail"
      x-swagger-router-controller: productDetail
      operationId: deleteAllProductDetail
      summary: Delete All Available productDetail.
      responses:
        '204':    # status code
          description: Default Response
  /productDetail/{productDetailId}:
    get:
      tags:
        - "ProductDetail"
      x-swagger-router-controller: productDetail
      operationId: getProductDetailById
      summary: Returns  productDetail.
      parameters:
        - name: productDetailId
          in: path
          description: productDetailId
          required: false
          schema:
            type: string
      responses:
        '200':    # status code
          description: Default Response
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/ProductDetail"
    patch: 
      x-swagger-router-controller: productDetail
      operationId: updateProductDetailById
      summary: update in a productDetail.
      tags:
        - "ProductDetail"      
      description: Register a new productDetail
      parameters:
        - name: productDetailId
          in: path
          description: productDetailId
          required: false
          schema:
            type: string      
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                productName:
                  type: string
                  example: Door
                productId:
                  type: string
                  example: sumit
                description:
                  type: string
                  example: goodquality
                category:
                  type: string
                  example: Doors
                price:
                  type: number
                  example: 233
                homeCategory:
                  type: string
                  example: bed
                subcategory:
                  type: string
                  example: goodquality
                availability:
                  type: boolean
                  example: true
                quantity:
                  type: number
                  example: 2
                rating:
                  type: number
                  example: 4
                images:
                  type: array
                  items:
                    type: string
                  example:
                    - bed
                    - sofa 
                assembly:
                  type: string
                  example: anything
                brand:
                  type: string
                  example: SARA
                material:
                  type: string
                  example: wooden
                color:
                  type: string
                  example: brown
                type:
                  type: string
                  example: bed
                totalOrder:
                  type: number
                  example: 5
                dimensionInInch:
                  type: string
                  example: 12in   
                dimensionInCem:
                  type: string
                  example: 2cm
                storage:
                  type: string
                  example: xyz
                weight:
                  type: string
                  example: 40KG.
                warranty:
                  type: number
                  example: 2
      responses: 
        '200':
          description: Created 
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"      
    delete:
      tags:
        - "ProductDetail"
      x-swagger-router-controller: productDetail
      operationId: deleteProductDetailById
      summary: Delete Available productDetailbyId.
      parameters:
        - name: productDetailId
          in: path
          description: productDetailId
          required: false
          schema:
            type: string        
      responses:
        '204':    # status code
          description: Default Response  
  /productDetail/ByProductID/{productId}:
    get:
      tags:
        - "ProductDetail"
      x-swagger-router-controller: productDetail
      operationId: getProductDetailByProductId
      summary: Returns  productDetail.
      parameters:
        - name: productId
          in: path
          description: productId
          required: false
          schema:
            type: string
      responses:
        '200':    # status code
          description: Default Response
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/ProductDetail" 
...
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: Authorization
      in: header
      description: Enter your Bearer token here
  schemas:
...
    ProductHome:
      type: object
      properties:
        productName:
          type: string
          example: Door      
        imageUrl:
          type: string
          example: doors.jpg 
        description:
          type: string
          example: good quality 
        productCategory:
          type: string
          example: Doors 
    ProductHomes:
      type: array
      items:
        $ref: "#/components/schemas/ProductHomes"                                                 
...
    ProductDetail:
      type: object
      properties:
                productName:
                  type: string
                  example: Door
                productId:
                  type: string
                  example: sumit
                description:
                  type: string
                  example: goodquality
                category:
                  type: string
                  example: Doors
                price:
                  type: number
                  example: 233
                homeCategory:
                  type: string
                  example: bed
                subcategory:
                  type: string
                  example: goodquality
                availability:
                  type: boolean
                  example: true
                quantity:
                  type: number
                  example: 2
                rating:
                  type: number
                  example: 4
                images:
                  type: array
                  items:
                    type: string
                  example:
                    - bed
                    - sofa 
                assembly:
                  type: string
                  example: anything
                brand:
                  type: string
                  example: SARA
                material:
                  type: string
                  example: wooden
                color:
                  type: string
                  example: brown
                type:
                  type: string
                  example: bed
                totalOrder:
                  type: number
                  example: 5
                dimensionInInch:
                  type: string
                  example: 12in   
                dimensionInCem:
                  type: string
                  example: 2cm
                storage:
                  type: string
                  example: xyz
                weight:
                  type: string
                  example: 40KG.
                warranty:
                  type: number
                  example: 2 
    ProductDetails:
      type: array
      items:
        $ref: "#/components/schemas/ProductDetail" 
...